+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Refreshed login page
 - Added Sara OAuth App configuration page button in the settings page
 - Added Sara OAuth App info/redirection links to login screen, create project page, and billing context create page
+- Starter prompts now auto submit
 
 ### Bug Fixes